+++ Aula01/exercicio01/src/Calculadora.java
@@ -1,15 +1,15 @@
 public class Calculadora {
 
-    public int soma(int valor1, int valor2){
+    public double soma(double valor1, double valor2){
         return valor1 + valor2;
     }
-    public  int subtracao(int valor1, int valor2){
+    public  double subtracao(double valor1, double valor2){
         return valor1 - valor2;
     }
-    public  float multplica(float valor1, float valor2){
+    public  double multplica(double valor1, double valor2){
         return  valor1 * valor2;
     }
-    public float divide(float valor1, float valor2){
+    public double divide(double valor1, double valor2){
         return valor1 / valor2;
     }
 }
